refactor(Table): client/配下をフラット配置に整理する
useTableHeadCellCountがclient/components配下からしか利用されておらず、
components/hooksへの下位分割自体が不要だったため、client/直下にフラットに
配置し直す。

componentsとhooksを分ける必要があるのはhookがclient component経由でない
独立した参照経路を持つ場合のみであり、SectioningContentのuseSectioningWrapper
のようなケースに限られる。この判断基準をCLAUDE.mdに追記する。
