--- packages/smarthr-ui/src/components/Table/client/hooks/useTableHeadCellCount/index.ts
@@ -1,1 +1,0 @@
-export { useTableHeadCellCount } from './useTableHeadCellCount'
