+++ packages/smarthr-ui/src/components/Table/index.ts
@@ -3,5 +3,5 @@ export { Th } from './Th'
 export { Td } from './Td'
 export { TdCheckbox } from './TdCheckbox'
 export { TdRadioButton } from './TdRadioButton'
-export { ThCheckbox, BulkActionRow, EmptyTableBody } from './client/components'
+export { ThCheckbox, BulkActionRow, EmptyTableBody } from './client'
 export { WakuWakuButton } from './WakuWakuButton'
