+++ packages/smarthr-ui/src/components/Table/stories/BulkActionRow.stories.tsx
@@ -1,5 +1,5 @@
 import { Table } from '../Table'
-import { BulkActionRow } from '../client/components'
+import { BulkActionRow } from '../client'
 
 import type { Meta, StoryObj } from '@storybook/react-webpack5'
