+++ packages/smarthr-ui/src/components/Table/stories/Table.stories.tsx
@@ -6,7 +6,7 @@ import { Td } from '../Td'
 import { TdCheckbox } from '../TdCheckbox'
 import { Th } from '../Th'
 import { WakuWakuButton } from '../WakuWakuButton'
-import { BulkActionRow, EmptyTableBody, ThCheckbox } from '../client/components'
+import { BulkActionRow, EmptyTableBody, ThCheckbox } from '../client'
 
 import type { Meta, StoryFn, StoryObj } from '@storybook/react-webpack5'
