+++ packages/smarthr-ui/src/components/Table/stories/VRTBulkActionRow.stories.tsx
@@ -1,6 +1,6 @@
 import { Table } from '../Table'
 import { Th } from '../Th'
-import { BulkActionRow } from '../client/components'
+import { BulkActionRow } from '../client'
 
 import type { Meta } from '@storybook/react-webpack5'
