+++ packages/smarthr-ui/src/components/Table/stories/VRTEmptyTableBody.stories.tsx
@@ -1,6 +1,6 @@
 import { Table } from '../Table'
 import { Th } from '../Th'
-import { EmptyTableBody } from '../client/components'
+import { EmptyTableBody } from '../client'
 
 import type { Meta } from '@storybook/react-webpack5'
